requirements.txt
camel-ai
streamlit
pandas
matplotlib
python-dotenv
TextBlob
praw